pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saga-story-system"
version = "0.1.0"
description = "S.A.G.A. — Story Analysis, Generation, and Archives."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4",
  "ebooklib",
  "google-genai",
  "mistralai",
  "pypdf",
  "PyMuPDF",
  "requests",
  "scikit-learn",
  "streamlit",
  "trafilatura",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
namespaces = true
include = [
  "analysis",
  "entities",
  "infrastructure",
  "prompts",
  "query",
  "rag",
  "services",
  "state",
  "timeline",
  "tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
